modules/inference_pipeline/simple-rag/pyproject.toml
[tool.poetry]
name = "simple-rag"
version = "0.1.0"
description = ""
authors = ["bharathsanthanam"]
readme = "README.md"
packages = [{include = "simple_rag"}]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.148"
python-dotenv = "^1.0.0"
unstructured = "^0.10.5"
pydantic = ">=2.7.4,<3.0.0"
awscli = { version = "^1.33.32", source = "PyPI" } 
langchain = "^0.2.11"
langchain-community = "^0.2.10"
pypdf = "^4.3.0"  # Corrected from pydpf2 to pypdf2
streamlit = "^1.37.0"
qdrant-client = "1.10.1"
fire = "^0.5.0"
transformers = "^4.33.1"
torch = {version = "2.0.1+cpu", source = "torch-cpu"}
pyyaml = "6.0.1"
certifi = "^2024.7.4"
faiss-cpu= "^1.8.0"
sentence-transformers = "^3.0.1"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.285"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"